This commit adds support to the bmi323 device on top of the pre-existing bmc150 kernel module.
Some new devices for example the ROG Ally and the Air Plus identify this chip in the ACPI table as a bmc150 so previously the original module was loaded,
but was erroring out as it cannot handle such device.

The device I own does not allow me to use the interrupt part of the device as the interrupt pin is not connected (or not advertised to be connected) hence
I avoided including on this commit anything related to IRQ.

This driver has already been proved to work well enough to be used in the switch emulator "yuzu".

While designing this module my main focus was not to alter the original driver and not to limit the original author in regard to future mofications,
and I was mostly able to achive this, except:
1) I added a new structure on top of the original one and added a field that is responsible for holding information
on what type of chip the module is currently managing
2) the previous point required the init function of the original driver to write that field in order to be sure no bmi323 code
was executed when the old part of the module is managing the device
3) as the original driver issued an i2c write on some register not really meant to be written in the bmi323 device I have made sure an i2c read to discover
the bmi323 is performed prior to that code: such read SHOULD fail in the older bmc150 IC for two reasons:
    - the i2c address is not reported in the memory map of the bmc150 in its datasheet
    - the i2c read attempts to get 4 bytes out of a 8-bit device
    - the fourth bit (the one that cannot be read from a bmc150 device) is initialized to 0 and bmi323 presence is signaled with a 1 in the LSB
      that is the fourth coming out of the device in temporal order