HACK: change TF API to fit current gamescope
upstream requested to identify if the type of transfer function a
property is capable to handle (basically EOTF or inv_EOTF). gamescope
understand it is implictly identified by the block/property position in
the pipeline and doesn't handle this difference yet. change the upstream
API to fit the current gamescope implementation.

Signed-off-by: Melissa Wen <mwen@igalia.com>